Validate requestId before clearing pending requests + dedupe type
Code review fixes:
- Add clearPendingRequestIfMatches helper that validates requestId matches
  the stored pending request before clearing (prevents clearing a newer
  request when responding to a stale one)
- Extract PendingInteractiveRequest to shared module to eliminate
  duplicate type definitions between frontend and backend
